# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/cash-flow-statement/{id}": {
      "get": {
        "summary": "Get cash flow statement by ID",
        "description": "Retrieve a specific cash flow statement record",
        "tags": [
          "Cash Flow Statement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved cash flow statement"
          },
          "404": {
            "description": "Cash flow statement not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update cash flow statement",
        "description": "Update an existing cash flow statement record",
        "tags": [
          "Cash Flow Statement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated"
          },
          "404": {
            "description": "Cash flow statement not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete cash flow statement",
        "description": "Delete a cash flow statement record",
        "tags": [
          "Cash Flow Statement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted"
          },
          "404": {
            "description": "Cash flow statement not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/cash-flow-statement": {
      "get": {
        "summary": "Get all cash flow statements",
        "description": "Retrieve all cash flow statement records with pagination",
        "tags": [
          "Cash Flow Statement"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 100
            },
            "description": "Maximum number of records to return"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of records to skip"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter by category"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved cash flow statements"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new cash flow statement",
        "description": "Create a new cash flow statement record",
        "tags": [
          "Cash Flow Statement"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "category"
                ],
                "properties": {
                  "category": {
                    "type": "string"
                  },
                  "net_income": {
                    "type": "number"
                  },
                  "changes_in_working_capital": {
                    "type": "number"
                  },
                  "depreciation_and_amortization": {
                    "type": "number"
                  },
                  "accounts_receivable": {
                    "type": "number"
                  },
                  "accounts_payable": {
                    "type": "number"
                  },
                  "deferred_taxes": {
                    "type": "number"
                  },
                  "other": {
                    "type": "number"
                  },
                  "cash_from_sale_of_capital_assets": {
                    "type": "number"
                  },
                  "cash_paid_for_purchase_of_capital_assets": {
                    "type": "number"
                  },
                  "increases_in_long_term_assets": {
                    "type": "number"
                  },
                  "proceeds_from_common_stock_issuance": {
                    "type": "number"
                  },
                  "proceeds_from_long_term_debt_issuance": {
                    "type": "number"
                  },
                  "dividends_paid_out": {
                    "type": "number"
                  },
                  "proceeds_from_preferred_stock_issuance": {
                    "type": "number"
                  },
                  "net_change_in_cash_and_cash_equivalents": {
                    "type": "number"
                  },
                  "cash_at_beginning_of_period": {
                    "type": "number"
                  },
                  "cash_at_end_of_period": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cash flow statement created successfully"
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}